Summary
This is a MySql/PHP Open Source (OS) application designed for the development of Alternate Reality Games (ARGs) based on Networked Narrative Storytelling (NNS).

It is part of the research project Principles of Liberty (PoL) conducted by Tabea Hirzel at the SMC University, Zug, Switzerland.

For further information on installation and usage of this software go to SPINDLE-WIKI.